+++ CHANGES.md
@@ -1,3 +1,4 @@
 # 0.7.0
 
 * First version of the SQLAlchemy EXASOL dialect released under BSD license
+* a lot of minor bug fixes to pass the SQLAlchemy test suite